fix: subscriber /mongodb connections are managed by the worker (#100)
Creating a subscriber/MongoDB connection should be outside the functions of defining a worker. As the primary idea of a worker is a process with the ability to startup and shutdown properly. The resources that might startup/shutdown with it are better run and defined by the user so the user doesn't bend to the will of a worker library